+++ pallets/voting/src/mock.rs
@@ -46,7 +46,7 @@ parameter_types! {
 	pub const MotionDuration: u64 = 2;
 	pub const MaxProposals: u32 = 100;
 	pub BlockWeights: frame_system::limits::BlockWeights =
-		frame_system::limits::BlockWeights::simple_max(Weight::from_ref_time(1024 as u64));
+		frame_system::limits::BlockWeights::simple_max(Weight::from_ref_time(1024_u64));
 }
 
 impl frame_system::Config for Test {
